% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ikunfrailty-package}
\alias{ikunfrailty-package}
\alias{ikunfrailty}
\title{ikunfrailty: imputation for kinship-induced frailty under nonrandom ascertainment}
\description{
\code{ikunfrailty} stands for Imputation of covariates for Kinship-induced
Unobserved frailty under Nonrandom ascertainment. It implements the proposed
posterior-draw multiple imputation (PDMI) engine developed for the V2.2
ascertainment-corrected frailtypack technical-note model. The package
derives the congenial imputation kernel from the user analysis formula and
the working covariate prior:
}
\details{
\deqn{
f(X_M \mid X_O,Y,P,A=1;\psi) \propto
\prod_i
\frac{s_i(Y_i,X_i)M_i(Y_i\mid X_i;\theta)}
{\alpha_i^F(X_i;\theta)}
g_i^A(X_i\mid P_i;\eta_X).
}

The validated first engine supports the scalar V2.2 setting: one continuous
PRS covariate and one binary carrier covariate, separately or jointly
missing, in an ascertainment-corrected Weibull/lognormal correlated frailty
model fit by \code{frailtypack::frailtyPenal()}.
}
\seealso{
\code{\link[=pdmi_frailty]{pdmi_frailty()}}, \code{\link[=pdmi_prior]{pdmi_prior()}}, \code{\link[=normal_kinship]{normal_kinship()}},
\code{\link[=carrier_hwe]{carrier_hwe()}}, \code{\link[=pen_summary]{pen_summary()}}, \code{\link[=pen_plot]{pen_plot()}}, \code{\link[=imputation_model]{imputation_model()}}
}
\author{
\strong{Maintainer}: Jiaqi Bi \email{jbi23@uwo.ca}

Authors:
\itemize{
  \item Jiaqi Bi \email{jbi23@uwo.ca}
  \item Osvaldo Espin-Garcia \email{oespinga@uwo.ca}
  \item Yun-Hee Choi \email{yun-hee.choi@schulich.uwo.ca}
}

Other contributors:
\itemize{
  \item Laurent Briollais \email{laurent@lunenfeld.ca} [contributor]
}

}
