% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{mvn_kinship}
\alias{mvn_kinship}
\title{Multivariate kinship prior placeholder}
\usage{
mvn_kinship(formula = ~1, covariance = "kinship+iid", ...)
}
\arguments{
\item{formula}{Mean formula.}

\item{covariance}{Covariance description.}

\item{...}{Reserved options.}
}
\value{
A \code{pdmi_mvn_kinship_prior} object.
}
\description{
Records a future multivariate continuous prior specification. The validated
V2.2 engine is scalar and will error if this prior is used in
\code{pdmi_frailty()}.
}
