% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pen_plot}
\alias{pen_plot}
\alias{pen_plot.pdmi_frailty}
\title{Plot penetrance curves}
\usage{
pen_plot(x, ...)

\method{pen_plot}{pdmi_frailty}(x, ci = x$pen_ci, conf.level = x$conf.level, ...)
}
\arguments{
\item{x}{A \code{pdmi_frailty} object.}

\item{...}{Unused.}

\item{ci}{If \code{TRUE}, draw confidence interval ribbons when available.}

\item{conf.level}{Confidence level used when intervals need to be added.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot penetrance curves
}
