% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pen_summary}
\alias{pen_summary}
\alias{pen_summary.pdmi_frailty}
\title{Summarize penetrance estimates}
\usage{
pen_summary(x, ...)

\method{pen_summary}{pdmi_frailty}(x, conf.level = x$conf.level, ci = x$pen_ci, ...)
}
\arguments{
\item{x}{A \code{pdmi_frailty} object.}

\item{...}{Unused.}

\item{conf.level}{Confidence level for intervals.}

\item{ci}{If \code{TRUE}, include \code{lower} and \code{upper}.}
}
\value{
A \code{pdmi_penetrance_summary} data frame.
}
\description{
Summarize penetrance estimates
}
