% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pdmi_frailty}
\alias{print.pdmi_frailty}
\title{Print a PDMI frailty fit}
\usage{
\method{print}{pdmi_frailty}(x, ...)
}
\arguments{
\item{x}{A \code{pdmi_frailty} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a PDMI frailty fit
}
