% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pdmi_frailty}
\alias{summary.pdmi_frailty}
\title{Summarize pooled PDMI parameter estimates}
\usage{
\method{summary}{pdmi_frailty}(object, conf.level = object$conf.level, ...)
}
\arguments{
\item{object}{A \code{pdmi_frailty} object.}

\item{conf.level}{Confidence level for Wald intervals.}

\item{...}{Unused.}
}
\value{
A \code{summary.pdmi_frailty} object.
}
\description{
Summarize pooled PDMI parameter estimates
}
